refactor: added router folder that performs route Fixes #20
