+++ packages/IoC/index.ts
@@ -1,8 +1,8 @@
 export * from "./scanners";
 export * from "./ContainerManager";
 export * from "./JSDocReader";
-export * from "./RouterExecutionContext";
-export * from "./GuardConsumer";
-export * from "./RouteParameterTransformer";
-export * from "./HeaderConsumer";
-export * from "./RenderConsumer";
+export * from "../router/RouterExecutionContext";
+export * from "../router/GuardConsumer";
+export * from "../router/RouteParameterTransformer";
+export * from "../router/HeaderConsumer";
+export * from "../router/RenderConsumer";
