+++ packages/example/views/login.hbs
@@ -7,7 +7,7 @@
     <body>
         <div>
             <h2>로그인</h2>
-            <form action="/login" method="post">
+            <form action="/auth/login" method="post">
                 <input type="text" name="username" placeholder="{{username}}" />
                 <input
                     type="password"
