+++ tsconfig.json
@@ -45,7 +45,9 @@
             "@stingerloom/factory": ["packages/factory"],
             "@stingerloom/factory/*": ["packages/factory/*"],
             "@stingerloom/services": ["packages/services"],
-            "@stingerloom/services/*": ["packages/services/*"]
+            "@stingerloom/services/*": ["packages/services/*"],
+            "@stingerloom/router": ["packages/router"],
+            "@stingerloom/router/*": ["packages/router/*"]
         } /* Specify a set of entries that re-map imports to additional lookup locations. */,
         // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
         // "typeRoots": [
